+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## QFeatures 0.99.3
 
+- New feature: the `rowData` method returns a list containing the 
+  `rowData` for all assays (see #86)
+  <2020-09-16 Wed>
 - Keep colnames when reading a single column assay (see #108)
   <2020-09-09 Wed>